Increase performance of Raster thread
All this does is avoid adding an animated container when the opacity is 0 (which it is most of the time).

The new line avoids an unintuitive behavior of the Flutter Framework, in which a fully transparent overlay cannot be ignored by the rasterizer. More info here: https://github.com/flutter/flutter/pull/72526#issuecomment-749185938.

I wasn’t able to test this, but I’m pretty sure this will have the effect of skipping the overlay animation when the opacity goes from 0 to non-zero (because AnimatedContainer only animates itself if it’s there for both the before and after — and I’m removing the before). If this is too jarring, we can think about other creative ways to still get that performance boost.